+++ cg/cli/demultiplex/sample_sheet.py
@@ -13,7 +13,7 @@ from cg.apps.demultiplex.sample_sheet.models import (
     FlowCellSampleDragen,
 )
 from cg.apps.demultiplex.sample_sheet.validate import get_sample_sheet_from_file
-from cg.apps.lims.samplesheet import flowcell_samples
+from cg.apps.lims.sample_sheet import flowcell_samples
 from cg.constants.constants import FileFormat
 from cg.constants.demultiplexing import OPTION_BCL_CONVERTER
 from cg.exc import FlowCellError
